Change `bbAppendn` argument type to `size_t`
Fixes lots of warnings about conversions changing value

+++ ncgen/bytebuffer.c
@@ -144,9 +144,9 @@ bbCatbuf(Bytebuffer* bb, const Bytebuffer* s)
 }
 
 int
-bbAppendn(Bytebuffer* bb, const void* elem, const unsigned int n0)
+bbAppendn(Bytebuffer* bb, const void* elem, const size_t n0)
 {
-  unsigned int n = n0;
+  size_t n = n0;
   if(bb == NULL || elem == NULL) return bbFail();
   if(n == 0) {n = strlen((char*)elem);}
   while(!bbNeed(bb,(n+1))) {if(!bbSetalloc(bb,0)) return bbFail();}
